Moved AtomicGetSetEphysMetadata to common_utils
